Docker image changes: Added a baseimage
Added a baseimage from which we make router and log-server images.
This base image has all the required packages for both the router
and the log-server. If any dependencies for either the router or
the log-server change, this baseimage ought to be updated. But it
makes deploying new code in docker containers quicker without
needing to download the packages from APT repositories everytime.

Usage is pretty much the same:

To setup, follow two steps:
$ make images   <= This creates baseimage and other docker images
$ make          <= This starts the docker containers

To cleanup, use any of the following (based on what you want)
$ make clean    <= This just terminates running containers
$ make clean-images     <= "clean" + cleanup of router, log-server image
$ make clean-all        <= "clean-images" + cleanup of baseimage