Updated python's machine-mon to get more data
Part of the infrastructure related visualization can be instrumenting
servers. Since we now have the capability to plot fairly generic
one-D time series data, this is going to be a very simple extension.

There is a definite general concern of affecting the measurements by
running a measurement tool. Which is why running it at fairly low
frequency is *highly* recommended.