New i3mote=>GDP bridge, based on libcoap
The earlier implementation was based on openwsn coap implementation,
which threw 'unknown message type' errors and would stop after a while.

`libcoap` seems to be a more complete implementation with support for
`OBSERVE`. Even though this is not a perfect implementation either,
(processes stop mysteriously after a while), we get around that by
restarting the processes if they die.