Moved GDPcache as part of package 'gdp'.
Since quite a few things in `gdp-if` use it, the hack of using GDPcache
as an app makes management difficult in the long-term. In general,
the python module ought to provide both high-level acess for quick-
start, as well as low-level access for more elaborate hacking.

Another such candidate to be migrated from an app to part of the package
is the key-value store. The argument for not doing it, however, is that
it is not a generic interface as GDPcache and requires the log to be
setup specifically in the key-value store format.