Removed automatically insert udata for async calls
As a workaround for the previous bug in the C library that didn't
handle the asynchronous events very well, python calls to the
C library functions automatically inserted an identifier for which
thread called the original function that resulted in the event
being generated in the first place. This is no longer needed, hence
removed.