Remove multiple py objs referring to same C mem
In the original design of the C library, if the same GCL was
opened multiple times, the returned data structures would point
to an existing allocated memory allocated earlier. This made for
very interesting consequences, in that the python bindings had
to manage their own mechanisms and workarounds (e.g. using __new__
instead of the normal __init__).

Since we have a better model at the C level now, such workarounds
are no longer needed. However, we still do need to maintain a
mapping of C pointers to corresponding python objects to ensure
that incoming events can be dispatched appropriately. [Hence the
leftover object_dir].