updated GDPcache.py for new API
GDPcache.py (and the Python API in general) currently has a
bug that resurfaces only for complex applications that rely
on garbage collection of objects.

Because there is no direct way for Python GC to know that an
event structure in C "contains" the datum (which "contains"
the buf, and so on), it will go ahead and cleanup the Python
event object when there are no references to the event object.
This has weird consequences for anyone holding on to a "datum"
object, as they will see no data whatsoever after GC does its
job.

A cleaner solution coming soon. But simple applications that
do not necessarily rely on GC should work.